refactor: added ContactsGroup component
Contacts are now fetched and rendered from ContactsGroup component instead of
the Contacts page.
