+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from './ContentBlock/ContentBlock';
 export * from './IconBox/IconBox';
 export * from './Contacts/Contacts';
 export * from './Alert/Alert';
+export * from './ContactsGroup/ContactsGroup';
