+++ src/layout/Solutions/Solutions.module.scss
@@ -3,26 +3,27 @@
 .Solutions {
     display: flex;
     align-items: center;
-    row-gap: 30px;
+    flex-flow: column wrap;
     justify-content: center;
-    flex-wrap: wrap;
-    margin-top: 30px;
-    flex-direction: column;
 
-    @media (min-width: 425px) {
-        flex-direction: row;
-        align-items: flex-start;
-        justify-content: space-around;
-        row-gap: 0;
-    }
+    margin-top: 30px;
+    row-gap: 30px;
 
     &__flexWrapper {
-        width: 100%;
-        max-width: 120px;
         display: flex;
         align-items: center;
-        justify-content: center;
         flex-direction: column;
+        justify-content: center;
+
+        width: 100%;
+        max-width: 120px;
         row-gap: 15px;
+    }
+
+    @media (width >= 425px) {
+        align-items: flex-start;
+        flex-direction: row;
+        justify-content: space-around;
+        row-gap: 0;
     }
 }
